Loading user avatars works correctly.
Makepad spits out TONS of errors related to failed PNG loads,
as we're always blindly attempting to load the avatar thumbnail as a PNG
and then trying to load it as a JPEG if the initial PNG attempt fails.

Instead, we need to check the image type before loading it.