Ensure the rooms list is always freshly updated.
Sometimes an `Event::Signal` gets lost (applied to a different event generation),
so we were missing them in the rooms_list event handler.
This manifested as some rooms failing to be displayed if they were discovered
after a user had clicked away from the rooms_list screen and then navigated
back to the screen.

Now, we check for updates to the rooms on every event handler,
which ensures we never miss any updates even if an `Event::Signal` was missed.