Support opening links via `robius-open`.
Special `matrix.to` links are excluded for now,
as those should be represented specially in the app itself.

Explicitly set the color of the status label at the bottom of
the rooms list, as the default Makepad theme has changed.

Depend on our own fork of Makepad, as the current `rik` branch
from the upstream repo has an infinite loop bug when rendering emoji.