Properly handle events when the user profile pane is shown
Events that flow into the RoomScreen either get forwarded to
the underlying timeline view or the user profile sliding pane.
If the pane is currently shown, mouse/touch/tap events are forwarded
to the pane *ONLY*, and not the underlying timeline.
Other events (e.g., draw events) are always forwarded to the underlying
timeline.
If the pane is not shown, events are also forwarded to the timeline view,
as normal.