subscribe to updates to the list of ignored users
This allows Robrix to be aware of the current logged-in user
ignoring or unignoring a user on a different client instance,
and respond to the fact that an ignore/unignore action has occurred
elsewhere, i.e., *beyond* the ignore user UI flow that is
driven from the UserProfileSlidingPane action buttons.

As such, we move the logic for re-paginating all cleared timelines
to the handler for updates to the ignored users lists.