Maintain our own list of ignored users
Do not attempt to trust or use the Matrix SDK's list of ignored users,
as it does not get updated properly in the current version of the SDK.

We now maintain our own list, which does properly get synced
when the logged-in users ignores/unignores other users on a different
client or session.
This includes correctly setting the `ignore_user_button`'s text.