fix the issue of message profiles/avatars not being properly written
There is a "bug" (maybe a bug) in Makepad where if you give multiple
DSL elements the same identifier name, then the widget reference you get
when querying it with `id!(...)` may be for a different widget
than the one you expect.

This commit renames the subwidgets of `ReplyPreview` in the DSL
in order to ensure they do not overlay with the same elements
that are part of a `Message` itself.

It also completely restructures the drawing logic for ReplyPreviews
to be done in a separate function and then to be called in the correct
place, i.e., when the item has not been cached.

The "fully-drawn" status of a Message's ReplyPreview is now factored in
to the conditional that determines whether a given message is
considered to be fully drawn (for the purpose of caching).