Refactor `ReplyPreview` into component views
This avoids the confusion of using the same single `ReplyPreview`
view for multiple purposes (an inline reply and a reply-to draft preview),
and instead clarifies the intended usage of each view
by using different views for those two purposes.

This also avoids the more complex logic of having to rememeber
to mark certain subviews as visible or not.