Move `TopSpace` out of `Timeline`'s PortalList and into `RoomScreen`
This vastly simplifies the code clarity and logic of dealing with
indices of PortalList items (`item_id`) vs. indices of Timeline items (`tl_idx`),
as they are now the same value.

The `TopSpace` is not yet shown, as that is blocked on issue #132.
Once that is addressed, we could show it as is, or animate it in and out
using a fixed `height` that the animator modifies, and/or make it an
overlay instead of a separate component that shifts the `Timeline` view down.