Ensure positional stability of the timeline in (almost) all cases
With this change, every visible item in the portal list is eligible
to be considered as a potential "anchor" item on which the viewport should
be fixed upon both before *and* after a background timeline update.

We iterate over all visible items and ensure that they have a valid
real unqiue event ID, have a valid known drawn position, and can be
found in the new items Vector.
This is a lot more flexible and powerful than the previous quick hacky
attempt that only considered the very first item in the viewport.

Utilize portal list functions that we added to makepad in order to
determine the relative position of every visible item in the portal list
with respect to the top of the portal list,
which we can then pass into `PortalList::set_first_id_and_scroll()`.