Save & restore replying-to preview state upon switching room timelines
Addresses a few to-do items in `RoomScreen` state management,
primarily the issue of a reply preview not actually being re-drawn
upon switching back to a previously-drawn room.

Also, the message input box is automatically given keyboard focus
after the user clicks the reply button for a message, making it
much easier for the user to start typing a reply.