Device verification is almost fully implemented. Untested.
A background async task runs a sync loop that adds event handlers
to the Client and then listens for incoming verification requests.
When it receives one asynchronously, it posts a `VerificationAction`
to the main UI thread, which is then handled by the top-level `App`.
The top-level `App` opens up and populates a `VerificationModal`,
which uses a tokio async channel that was sent along with the
`VerificationAction` to send user interactive responses back to the
async task.
The background async task uses those responses to communicate with
to the other device participating in the verificaiton via
standard Matrix SDK calls.
So basically, there are two state machines that get kept in lockstep
based on user-side actions and SDK-side actions,
and the async task and main UI thread communicate with each other
by posting actions to the main UI thread and then sending responses
back to the async task over a tokio channel.

Other changes:
* Consolidate Robrix-styled icon buttons into a separate widget,
  extracting it out of `user_profile`rs.
* Add a checkmark SVG icon to be used for Ok/Accept buttons.