Ensure that the location manager is initialized on the main UI thread.
Apparently, apple platforms (and likely others) require that the location manager
run (or at least be initialized)  on a thread that has  "run loop",
which the main thread is guaranteed to have.
See: <https://stackoverflow.com/questions/25916841/cllocationmanager-startupdatinglocation-not-calling-locationmanagerdidupdateloc>

We could technically init the location manager on a background thread
as long as we execute a run loop for that thread,
but that may not work on other platforms, and is a needless complication
when we already have a dedicated thread that has a run loop.

Unfortunately, we're still getting an `AuthorizationDenied` error when attempting
to receive a location update, so we need to move over to packaging Robrix
into a proper app bundle such that we can specify both the
`entitlements.plist` and `Info.plist` files that get bundled into the .app/.dmg.