Location sharing to a Matrix room now works!
Currently only tested on macOS, and the application must be fully
packaged and codesigned into a valid app bundle, and then installed
to the `/Applications` folder on macOS in order to work.
You also have to manually allow Robrix to access location in:
System Preferences -> Privacy & Security -> Location Services.

Location messages are not yet actually displayed correctly in
the timeline room screen view, but that's an easy fix.