Completely redesign pagination code, support backwards pagination upon scroll
Remove the separate pagination state subscriber task, which was previously
being spawned once for every room. That was unnecessary and actually provided
less information than just waiting for the result of the original task
that actually called the paginate function.

Store the most-recently-scrolled-through item index in `TimelineUiState`
such that we can determine whether when the user is scrolling upwards
and has just hit the top-most item, at which point we fire off
a backwards pagination request.