Only paginate automatically the first time a room is opened
Clean up various pagination code.

Re-format the TopSpace (for pagination loading messages) to be
a lightly-transparent green, and to be an Overlay atop the timeline
such that it doesn't affect the positional stability of timeline items.

Refactor RoomScreen event-handling functions to remove redundant calls
to `self.portal_list(id!(timeline.list))`. We now call this once
at the top of the `RoomScreen::handle_event()` in order to have a
single `PortalListRef` that can be used across all the related
event handling functions.