Optimize common diff pair of Remove + Insert as an in-place update
This is much more efficient as it doesn't result in running all of
the cleanup code required to remove a room. This happens all the time
in the new room_list_service, but it's never used to actually remove
a room -- just to modify its positional order as a form of "sorting"
the rooms list, which Robrix as a client does not care about.