Finished basic infra for paginating backwards to a certain event.
A modal is displayed while the backwards pagination requests are
ongoing, and the user can cancel/abort the search if it's taking too long.

Tested to work for basic scenarios; probably still needs more testing.
For example, I haven't tested the cancel feature yet, or the cleanup
that happens if you close a RoomScreen while the modal is being displayed
and the backwards pagination procedure is running in the background.