Use our improved `PortalList::smooth_scroll_to()` behavior
This changeset ensures that we don't scroll through potentially
hundreds or thousands of items on the way to a much older message.
There is now a maximum configurable window (number of items) that
will be included in the scrolling animation, such that we don't get
bogged down rendering thousands of items unnecessarily that barely
get shown during the animation.