Use `SelectedRoom` everywhere in the main content views
This ensures that room names are always preserved and visible
when switching between Desktop and Mobile-sized UI layouts
via the AdaptiveView.

Ideally this shouldn't be necessary, and we should only ever need
to store `OwnedRoomId`s in all of the `MainDesktopUi` fields.
I also think this change might be eventually problematic because
it'll be difficult to properly propagate a changed room name
through all of the app-wide states.