Update to latest makepad with new implicit redraw API (WIP)
This builds and runs, but the performance is not great.

The biggest problem is that the inner RoomScreen content
(within a Tab in the Dock) no longer refreshes itself
when the user selects a room from the RoomsList.
If the user manually re-selects the tab it will refresh properly
this happens both when selecting a new room and when closing
a tab to reveal the prior room.

Another (smaller) problem is that the scrolling of the RoomsList
is mucH slower and choppier than before.