Move message context menu to the top-level `app.rs`
This enables us to show only a single context menu across
the entire app, and also for it to spill over the bounds
of a single RoomScreen view, which is exactly how most users
will expect a standard context menu to behave.

We still need to use the dimensions of the context menu
to ensure that it doesn't render partially off-window.

We also need to fix up the top-level App's event handler
such that "interactive" events are not passed down to
the underlying views if the context menu is currently being shown,
and instead pass those events directly to the context menu only.
This should be the same approach we use with overlay "pane"
widgets in the RoomScreen.

The context menu actions are still not implemented.