Use the `hits()` API for hover in/out events in `Message`
This is not only more efficient in terms of redraws,
but also much simpler to understand.

Now, you can also long-press on the replied-to message preview area
above a message (or right-click on it) to bring up the context menu,
which is more ergonomic, especially when a message is really short
and hard to press with your finger.

Additionally, we only jump to replied-to messages upon a primary hit,
not upon any hit.