Use our new hit handling functionality in Makepad
The general idea is to let child widgets handle a given event
before their parent widget handles it.
If a given child widget handles that event, the parent widget
won't handle it duplicately.
However, if no child widgets handle it, then the event will be able
to be handled by the parent widget.

Now, a child widget can handle an event and then mark it as not-handled,
meaning that another widget will *also* be able to handle it.

Note that this Makepad change still preserves the historical behavior
of an event being marked as handled by default upon *any* call to
`event.hits()`, so if you want to override that behavior, you have to
explicitly call `EventHandled::mark_as_handled(false)` for that specific
`Hit` variant.

This also removes the our use on the previous callback-based implementation
of `hits_without_mark_as_handled()`, which was a lot harder to use
and less flexible.

See https://github.com/makepad/makepad/pull/673 for more details.