Adopt new hit-handling logic. Fix clicks on reply previews.
For things like `RoomPreview` and the `replied_to_messsage` above
a given `Message` body, we handle hits on those widgets *before* we
allow other child widgets to handle hits/events on themselves.

This ensure that sub-widgets within a `RoomPreview` or a
`replied_to_message` view *cannot* handle a click (such a link, avatar),
and that a click on one of those "parent" widgets just acts to
select it and not interact with subwidgets within it.