Redesign SearchBar into a single shared RoomFilterInputBar
The new `RoomFilterInputBar` is shared across both the
narrow "Mobile"-oriented UI and the wider "Desktop"-oriented UI,
and is now shown in different places on each.

We place the filter bar inside of a `CachedWidget` to ensure that
there is only a singleton instance of the `RoomFilterInputBar`
globally across the whole app.
This means that the text is now shared and that there is no need
to filter its actions across multiple different SearchBar actions.

Thus, it now automatically works when transitioning between the mobile
and desktop views.