Use a common Timestamp widget for all events
The new timestamp widget now shows a custom tooltip upon hover
with the full date/time info, so we can now also remove the
datestamp beneath the timestamp for non-condensed messages.

TODO: timestamps aren't horizontally aligned in all event types.
The SmallStateEvent's timestamps are slightly to the right of
the timestamps displaying for a message.