Properly save/restore the state of the EditingPane
Previously, we were re-using the `show_editing_pane()` function
to act as a "restore state" function, but that has some issues.
Now we have a dedicated save/restore function pair, and we also
make sure to reset the state of the EditingPane if there was no
pre-existing saved state to be restored.