Ensure that a room is loaded before issuing first-time actions
Such as requesting back pagination, subscribing to typing notices,
fetching the room member list, subscribing to the current user's read receipts,
getting room power levels for the current user, etc.

Make `set_displayed_room()` the only code path in which room details
are populated, even in the case of deferred loading after restoring the
state of the UI from a prior execution of Robrix.
Ensuring this "orthogonality" simplifies the code, especially that which
deals with whether a given room has been loaded and then transitioning
from the non-loaded state to the fully-loaded and fully-displayed state.

General clean-up after merging in better support for mentioning users.