Use DSL namespace linking to cfg-in/out TSP-specific widgets
This approach collects all TSP-specific widgets that are used by
non-TSP widgets (those that lie at the boundary between TSP and
non-TSP UI DSL code) and then places them under either the
* "tsp_enabled" namespace, for real TSP widgets that are included
  in the build when the `tsp` feature is enabled, or
* "tsp_disabled" namepsace, for dummy TSP widgets that are just
  empty placeholders for use when the `tsp` feature is *not* enabled.

In app startup, we create a new DSL namespace "tsp_link", and then
link that to either one of those namespaces, based on whether or not
the `tsp` feature is enabled.

This approach works, but still requires separate namespace handling
on a per-namespace (per-cargo feature) basis, which might get tedious
when more and more features are introduced.