Add infra for verifying other users via TSP
Things are largely untested, but the basic UI and logic are there.

TODO: still need to add a UI pane for accepting/rejecting an incoming
TSP verification request (similar to the VerificationModal, but for
TSP instead of Matrix-native verification requests).

Added a TSP section to the UserProfileSlidingPane that allows the user
to enter a DID and associate it with another Matrix user ID,
upon which a verification request will be sent to them.
(This is untested)

Fix issue where interactive events are delivered to the overlay views
in the RoomScreen (UserProfileSlidingPane, LoadingPane).