Fix joining replacement room and ensure navigation works
Previously, PR #536 added support for showing a tombstoned room
and information about its successor/replacement room,
but did not fully handle the case where a new room is unknown.
We now handle that properly, and also support deferred navigation
to the newly-joined room that waits for it to actually joined and
fully loaded/received by our client from the homeserver,
at which point it can actually jump to that new room.