Ensure that message details are only set for a drawn item once
Recently, a minor bug snuck in where it was possible for a message item's
widget to have incomplete `MessageDetails`, because the related event ID
was not always calculated but was still being re-set for the message item
when it did not need to be re-set.

Now we only set it once when the message *content* is first drawn,
which is correct in all cases, and also more efficient!