Ensure that the RoomInputBar components don't exceed 5/8 of the RoomScreen's height
This primarily applies to the TextInput within the RoomInputBar,
but it also limits the height of everything else within the RoomInputBar,
such as the ReplyPreview + TextInput together.

Update makepad version to include Eddy's latest commit that supports
Fit size bounds with maximum sizes that are relative to an ancestor widget.