Handle all changes in room info in `update_room()`
Instead of checking the latest event's content to catch changes,
(which is fallible as it may miss certain events),
we now handle it entirely in the stream of changes that comes from
the RoomListService.

I still feel like there must be a better way to do this, since we ideally
wouldn't need to maintain virtually the entire state of the room in a
separate cached struct (see `RoomListServiceRoomInfo`),
but I currently cannot see any other means of doing so.

We also still have to check for an update to the latest event
ON EVERY CHANGE to the room, which seems wildly inefficient.
We can't even use the `new_latest_event_timestamp()` to check if the
latest event actually has been updated more recently than others.
One reason for this is that certain events, like redactions, seem to
not update the timestamp, so if we don't blindly update the latest event
upon every room change, then redactions are skipped (incorrectly).

However, we have still improved the handling of lots of other room info
kinds, e.g., tags, unreads, room display name & avatar, tombstoned state,
`is_direct` state, power levels, etc.