Fetch full room previews dynamically (e.g., for successor rooms)
Instead of only looking through our client's local rooms list,
we now use `Client::get_room_preview()` to fetch the full details of
an unknown room from the homeserver.

Currently this is only used when showing the successor room for a
tombstoned room (as part of the `TombstoneFooter`), but it can also be
used for future UI views that show some basic info about a room
that the user is interested in joining or jumping to.

Also, this fixes a minor display issue in the `RoomInputBar` in which
the user power levels were not used to show/hide select subviews
when restoring the state of that room's input bar.